requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.11
fake-useragent==0.1.11
idna==3.3
mysql-connector-python==8.0.28
protobuf==3.19.4
requests==2.27.1
soupsieve==2.3.1
urllib3==1.26.8
useragent==0.1.1
